requirements.txt
requests>=2.31.0
feedparser>=6.0.11
pandas>=2.1.0
python-dateutil>=2.8.2
